+++ pkg/hw/display/displays_rpi.go
@@ -12,7 +12,7 @@ import (
 
 func (this *Displays) Enumerate() []gopi.Display {
 	displays := make([]gopi.Display, 0, rpi.TV_MAX_ATTACHED_DISPLAYS)
-	for i := uint16(0); i < rpi.TV_MAX_ATTACHED_DISPLAYS; i++ {
+	for i := uint32(0); i < rpi.TV_MAX_ATTACHED_DISPLAYS; i++ {
 		if display, err := this.Open(i); err == nil {
 			displays = append(displays, display)
 		}
